requirements.txt
asgiref==3.7.2
Django==5.0.1
django-admin-thumbnails==0.2.8
django-ckeditor==6.7.0
django-js-asset==2.2.0
django-mptt==0.16.0
pillow==10.2.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
whitenoise==6.6.0
